Avatar generator components split up in jsx files
