+++ app/javascript/avatar_generator/index.js
@@ -1,0 +1,1 @@
+import AvatarDrawingDisplayShirt from './AvatarDrawingDisplayShirt.jsx';
